refactor(yukigo): separate primitives in diff files
