+++ packages/yukigo/tests/interpreter/PatternSystem.spec.ts
@@ -16,7 +16,7 @@ import { createStream } from "../../src/interpreter/utils.js";
 import { RuntimeContext } from "../../src/interpreter/components/RuntimeContext.js";
 import { YukigoKernel } from "../../src/interpreter/components/kernel/index.js";
 import { InterpreterVisitor } from "../../src/interpreter/components/Visitor.js";
-import { LazyList } from "../../src/interpreter/runtime.js";
+import { LazyList } from "../../src/primitives/LazyList.js";
 
 const s = (v: string) => new SymbolPrimitive(v);
 const n = (v: number) => new NumberPrimitive(v);
